+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Fixed
+
+- Fixed a crash when the commit message was empty
+- Fixed newline formatting. Now the trailer is correctly preceded by an empty line
+  instead of having two trailing empty lines.
+
 ## [1.0.0] - 2025-05-21
 
 ### Added